Eliminate agency/execution/detail/execution_policy_traits.hpp
