+++ agency/detail/control_structures/is_bulk_call_possible_via_execution_policy.hpp
@@ -2,7 +2,7 @@
 
 #include <agency/detail/config.hpp>
 #include <agency/detail/is_call_possible.hpp>
-#include <agency/execution/detail/execution_policy_traits.hpp>
+#include <agency/execution/execution_policy/execution_policy_traits.hpp>
 #include <type_traits>
 
 namespace agency
